patch: split up app god file
